+++ CodingCat.RabbitMq/CodingCat.RabbitMq/Interfaces/IExchangeProperty.cs
@@ -1,7 +1,0 @@
-﻿namespace CodingCat.RabbitMq.Interfaces
-{
-    public interface IExchangeProperty : IDeclarableProperty
-    {
-        string Type { get; }
-    }
-}