+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+#### 0.3.0
+
+- removed project `PubSub`
+- migrated to `CodingCat.RabbitMq.Abstractions` 0.1
+- enhanced for using dependency injection
+
+
 #### 0.2.0
 
 fixed #1 [https://github.com/CodingCatHongKong/dotnet-CodingCat.RabbitMq/issues/1]